tests(helpers): separate grpc functions (#13552)
It is one of the serial refactors of helpers.lua

KAG-5242
